feat(auth): add support for OAuth2 Client Credentials
Authenticating using a bare OAuth2 Access Token is not very convenient
since it has limited validity and needs to be aquired by other means
outside of the provider.

Bitbucket supports authentication via Client Credentials Grant, which is
easier to use.

This change adds support for the `oauth_client_id` and
`oauth_client_secret` provider configuration fields (and environment
variables) and updates the provider documentation.

This is an alternative to the existing auth via `username`/`password` or
`oauth_token`.